core: Extract OpenGL glow pipeline into reusable class.
Move the DMD glow rendering (shaders, framebuffers, blur passes
and GL function bindings) out of NativeOpenGlRenderer into a new
VirtualDmdOpenGlPipeline class. NativeOpenGlRenderer now owns only
the window context and delegates frame rendering to the pipeline,
making the glow pipeline reusable across renderers.
